Move utility function to utility function file
