+++ src/scripts/delay_posting.js
@@ -1,44 +1,9 @@
-import { keyToCss } from '../util/css_map.js';
-import { inject } from '../util/inject.js';
 import { registerPostOption, unregisterPostOption } from '../util/post_actions.js';
 import { getPreferences } from '../util/preferences.js';
+import { editPostFormStatus } from '../util/react_props.js';
 
 const symbolId = 'ri-timer-flash-line';
 let delayMs;
-
-const controlPostFormStatus = (status, publishOn) => {
-  const button = document.currentScript.parentElement;
-  const reactKey = Object.keys(document.currentScript.parentElement).find(key => key.startsWith('__reactFiber'));
-
-  const isScheduled = status === 'scheduled';
-  let fiber = button[reactKey];
-  while (fiber !== null) {
-    if (fiber.stateNode?.state?.isDatePickerVisible !== undefined) {
-      fiber.stateNode.setState({ isDatePickerVisible: isScheduled });
-      break;
-    } else {
-      fiber = fiber.return;
-    }
-  }
-
-  fiber = button[reactKey];
-  while (fiber !== null) {
-    if (fiber.stateNode?.setFormPostStatus && fiber.stateNode?.onChangePublishOnValue) {
-      fiber.stateNode.setFormPostStatus(status);
-      if (publishOn) fiber.stateNode.onChangePublishOnValue(new Date(publishOn));
-      break;
-    } else {
-      fiber = fiber.return;
-    }
-  }
-};
-
-const editPostFormStatus = (status, publishOn) => {
-  const button = document.querySelector(`${keyToCss('postFormButton')} button`);
-  if (!button) throw new Error('Missing button element to edit post form status');
-
-  inject(controlPostFormStatus, [status, publishOn], button);
-};
 
 export const main = async function () {
   const { delaySeconds } = await getPreferences('delay_posting');
@@ -46,7 +11,7 @@ export const main = async function () {
 
   registerPostOption('delay-posting', {
     symbolId,
-    onclick: () => editPostFormStatus('scheduled', Date.now() + delayMs)
+    onclick: () => editPostFormStatus('scheduled', new Date(Date.now() + delayMs))
   });
 };
